refactor: ♻️ explode utils into main application
