+++ apps/web/package.json
@@ -17,7 +17,6 @@
   "dependencies": {
     "@hookform/resolvers": "5.1.1",
     "@popcorn.fyi/api-clients": "workspace:*",
-    "@popcorn.fyi/utils": "workspace:*",
     "@tanstack/match-sorter-utils": "8.19.4",
     "@tanstack/react-query": "5.81.5",
     "@tanstack/react-router": "1.122.0",
