+++ apps/web/src/components/table/year-table-cell.tsx
@@ -1,6 +1,6 @@
 import type { CellContext } from "@tanstack/react-table";
 
-import { year } from "@popcorn.fyi/utils";
+import { year } from "@/lib/year";
 
 export const YearTableCell = (
   props: CellContext<
