+++ apps/web/src/routes/_layout/movies.$id.index.tsx
@@ -1,10 +1,12 @@
-import { currency, date, time } from "@popcorn.fyi/utils";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
 
 import { movieDetailsOptions } from "@/api/movie/details";
 import { MediaOverviewList } from "@/components/media/media-overview-list";
 import { Prose } from "@/components/shared/prose";
+import { currency } from "@/lib/currency";
+import { date } from "@/lib/date";
+import { time } from "@/lib/time";
 
 export const Route = createFileRoute("/_layout/movies/$id/")({
   component: RouteComponent,
