--- libs/utils/README.md
@@ -1,1 +1,0 @@
-# @popcorn.fyi/utils
