--- libs/utils/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@popcorn.fyi/utils",
-  "private": true,
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "coverage": "vitest run --coverage",
-    "lint": "eslint --cache .",
-    "test": "vitest",
-    "typecheck": "tsc"
-  },
-  "devDependencies": {
-    "@popcorn.fyi/eslint": "workspace:*",
-    "@popcorn.fyi/ts": "workspace:*"
-  }
-}
