--- libs/utils/src/index.ts
@@ -1,6 +1,0 @@
-export { currency } from "./currency";
-export { date } from "./date";
-export { limit } from "./limit";
-export { shuffle } from "./shuffle";
-export { time } from "./time";
-export { year } from "./year";
