--- libs/utils/vitest.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    coverage: {
-      reporter: ["text", "json"],
-    },
-    globals: true,
-  },
-});
